search-ms/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
shapely==2.0.6
quads==1.1.0
